+++ compare.h
@@ -21,5 +21,5 @@ void compare_continue(scanstate *ss, const char *ptr, int len);
 void compare_end(scanstate *cmp);
 
 // no better place to put this for now...
-char* substitute_string(pcrs_job *job, const char *cp, const char *ce, int *newsize);
+char* substitute_string(pcrs_job *job, const char *cp, const char *ce, size_t *newsize);
